The unit test of RingBuffer was executed, and some problems were resolved.
git-svn-id: https://tmp.openrtm.org/svn/OpenRTM-aist-Java/branches/RELENG_1_0_0@119 9a6ac7ee-f497-465d-b832-7029f4913c1d

+++ jp.go.aist.rtm.RTC/src/jp/go/aist/rtm/RTC/buffer/BufferBase.java
@@ -120,7 +120,7 @@ public interface BufferBase<DataType> {
      */
     public ReturnCode write(final DataType value);
     public ReturnCode write(final DataType value,
-                             long sec, long nsec);
+                            int sec, int nsec);
 
 
     /**
