+++ jp.go.aist.rtm.RTC/src/jp/go/aist/rtm/RTC/port/publisher/PublisherFlush.java
@@ -104,7 +104,7 @@ public class PublisherFlush extends PublisherBase implements ObjectCreator<Publi
      * @param usec
      * @return ReturnCode
      */
-    public ReturnCode write(final OutputStream data, long sec, long usec) {
+    public ReturnCode write(final OutputStream data, int sec, int usec) {
         if (m_consumer == null ) { return ReturnCode.PRECONDITION_NOT_MET; }
 
         return m_consumer.put(data);
