+++ jp.go.aist.rtm.RTC/src/jp/go/aist/rtm/RTC/port/publisher/PublisherPeriodic.java
@@ -406,7 +406,7 @@ public class PublisherPeriodic extends PublisherBase implements Runnable, Object
      * @param usec
      * @return ReturnCode
      */
-    public ReturnCode write(final OutputStream data, long sec, long usec) {
+    public ReturnCode write(final OutputStream data, int sec, int usec) {
         rtcout.println(rtcout.PARANOID, "write()" );
         if (m_consumer == null) { return ReturnCode.PRECONDITION_NOT_MET; }
         if (m_buffer == null) { return ReturnCode.PRECONDITION_NOT_MET; }
